+++ src/app/product/componens/product-list/product-list.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ProductsService } from '../../services/products.service';
 import { IProduct } from '../../../shared/types';
+import { SortService } from '../../../shared/sort.service';
 
 @Component({
   selector: 'app-product-list',
@@ -8,12 +9,16 @@ import { IProduct } from '../../../shared/types';
   styleUrls: ['./product-list.component.scss']
 })
 export class ProductListComponent implements OnInit {
-  products: IProduct[] | undefined
+  products!: IProduct[];
 
-  constructor(public productService: ProductsService) { }
+  constructor(
+    public productService: ProductsService,
+    public sortService: SortService
+  ) { }
 
   ngOnInit(): void {
     this.products = this.productService.getProducts();
+    // this.sortService.selectedFilter$.subscribe(res => console.log(res))
   }
 
   onAddToCart($event: any): void {