+++ src/app/shared/types.ts
@@ -4,18 +4,22 @@ export interface IProduct {
   title: string;
   price: number;
   description: string;
-  category: any;
+  category: TCategory;
   image: string;
-  rating: IRating;
-  isAvailable: boolean
-}
-
-export interface IRating {
-  rate: number;
+  rating: number;
   count: number;
+  isAvailable: boolean;
 }
 
 export interface ICartProduct extends IProduct {
   quantity: number;
 }
-export type TCategory =  "mensClothing" | "jewelery" | "electronics" | "womensClothing"
+export type TCategory = "mensClothing" | "jewelery" | "electronics" | "womensClothing"
+
+export interface ISearchValue {
+  filter: TFilterNames,
+  sort: boolean
+}
+
+export type TFilterNames = 'price' | 'rating' | 'category'
+export type TFilterNamesCart = 'price' | 'rating' | 'category'| 'quantity'